SchemeRegistry: add method to check if a scheme is secure
URLLineEdit: improve method of coloring in a url based on the scheme,
host and path. Also be more tolerant in treating any user input as a url
that may be colored in.

Navigation bar: changed border color of url line edit + search line
edit to a slightly less dark grey